fix: organize tests to reflect src
