+++ src/benchq/magic_state_distillation/autoccz_factories.py
@@ -1,7 +1,7 @@
 import math
 from typing import Iterator, Tuple
 
-from ..quantum_hardware_modeling.gidney_surface_code_model import (  # noqa: E501
+from ..quantum_hardware_modeling.fowler_surface_code_model import (  # noqa: E501
     get_total_logical_failure_rate,
     physical_qubits_per_logical_qubit,
 )
