+++ src/benchq/problem_embeddings/__init__.py
@@ -3,3 +3,4 @@
 ################################################################################
 from ._trotter import get_trotter_circuit, get_trotter_program
 from .qsp._qsp import get_qsp_circuit, get_qsp_program
+from .quantum_program import QuantumProgram, get_program_from_circuit
