+++ src/benchq/problem_embeddings/_qaoa.py
@@ -7,9 +7,7 @@ from orquestra.quantum.circuits import Circuit
 from orquestra.quantum.operators import PauliSum
 from orquestra.vqa.algorithms.qaoa import QAOA
 
-from ..algorithms.data_structures import (
-    AlgorithmImplementation,
-    ErrorBudget,
+from .quantum_program import (
     QuantumProgram,
 )
 
@@ -52,12 +50,3 @@ def get_qaoa_program(hamiltonian: PauliSum, n_layers: int = 1) -> QuantumProgram
     random_params = np.random.uniform(-np.pi, np.pi, 2)
     circuit = qaoa.get_circuit(random_params)
     return QuantumProgram([circuit], n_layers, lambda x: [0] * x)
-
-
-def qaoa_algorithm(
-    hamiltonian: PauliSum, n_layers: int = 1, failure_tolerance: float = 1e-3
-):
-    program = get_qaoa_program(hamiltonian, n_layers)
-    return AlgorithmImplementation(
-        program, ErrorBudget.from_even_split(failure_tolerance), 1
-    )
