+++ src/benchq/problem_ingestion/time_marching_matrix_properties.py
