+++ src/benchq/resource_estimators/default_estimators.py
@@ -1,5 +1,5 @@
 from functools import partial
-from typing import Any, List, Optional
+from typing import List, Optional
 
 import numpy as np
