+++ tests/benchq/problem_embeddings/time_marching/test_compression_gadget.py
